feat: project deploy commit diff add e2e test
