+++ plugins/refdiff/e2e/raw_tables/cicd_pipeline_commits.csv
@@ -1,0 +1,6 @@
+pipeline_id,created_at,updated_at,_raw_data_params,_raw_data_table,_raw_data_id,_raw_data_remark,commit_sha,branch,repo_id,repo_url
+github:GithubRun:1:1,2022-11-17 06:35:35.758,2022-11-17 06:35:35.758,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1496,"",commit_sha1,main,github:GithubRepo:1:484251804,""
+github:GithubRun:1:2,2022-11-17 06:35:35.758,2022-11-17 06:35:35.758,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1495,"",commit_sha2,main,github:GithubRepo:1:484251804,""
+github:GithubRun:1:3,2022-11-17 06:35:35.758,2022-11-17 06:35:35.758,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1494,"",commit_sha3,fix-edit-url,github:GithubRepo:1:484251804,""
+github:GithubRun:1:4,2022-11-17 06:35:35.758,2022-11-17 06:35:35.758,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1493,"",commit_sha4,main,github:GithubRepo:1:484251804,""
+github:GithubRun:1:5,2022-11-17 06:35:35.758,2022-11-17 06:35:35.758,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1492,"",commit_sha5,main,github:GithubRepo:1:484251804,""
