+++ plugins/refdiff/e2e/raw_tables/cicd_pipelines.csv
@@ -1,0 +1,6 @@
+id,created_at,updated_at,_raw_data_params,_raw_data_table,_raw_data_id,_raw_data_remark,name,result,status,type,duration_sec,created_date,finished_date,environment,cicd_scope_id
+github:GithubRun:1:1,2022-11-17 06:35:35.708,2022-11-17 06:35:35.708,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1496,"",Production workflow,SUCCESS,DONE,"",491,2022-04-22 03:46:49,2022-04-22 03:55:00,"",github:GithubRepo:1:484251804
+github:GithubRun:1:2,2022-11-17 06:35:35.708,2022-11-17 06:35:35.708,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1495,"",Production workflow,SUCCESS,DONE,"",236,2022-04-22 05:00:38,2022-04-22 05:04:34,"",github:GithubRepo:1:484251804
+github:GithubRun:1:3,2022-11-17 06:35:35.708,2022-11-17 06:35:35.708,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1494,"",Build workflow,SUCCESS,DONE,"",102,2022-04-22 05:58:18,2022-04-22 06:00:00,"",github:GithubRepo:1:484251804
+github:GithubRun:1:4,2022-11-17 06:35:35.708,2022-11-17 06:35:35.708,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1493,"",Production workflow,SUCCESS,DONE,"",201,2022-04-22 06:39:34,2022-04-22 06:42:55,"",github:GithubRepo:1:484251804
+github:GithubRun:1:5,2022-11-17 06:35:35.708,2022-11-17 06:35:35.708,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_runs,1492,"",Production workflow,SUCCESS,DONE,"",236,2022-04-22 06:55:58,2022-04-22 06:59:54,"",github:GithubRepo:1:484251804
