+++ plugins/refdiff/e2e/raw_tables/cicd_tasks.csv
@@ -1,0 +1,5 @@
+id,created_at,updated_at,_raw_data_params,_raw_data_table,_raw_data_id,_raw_data_remark,name,pipeline_id,result,status,type,duration_sec,started_date,finished_date,environment,cicd_scope_id
+github:GithubJob:1:1,2022-11-17 06:37:11.409,2022-11-17 06:37:45.836,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_jobs,748,"",build-and-deploy,github:GithubRun:1:1,SUCCESS,DONE,DEPLOYMENT,480,2022-04-22 03:46:58,2022-04-22 03:54:58,PRODUCTION,""
+github:GithubJob:1:2,2022-11-17 06:37:11.409,2022-11-17 06:37:45.836,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_jobs,749,"",build-and-deploy,github:GithubRun:1:2,SUCCESS,DONE,DEPLOYMENT,225,2022-04-22 05:00:47,2022-04-22 05:04:32,PRODUCTION,""
+github:GithubJob:1:3,2022-11-17 06:37:11.409,2022-11-17 06:37:45.836,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_jobs,750,"",build,github:GithubRun:1:3,SUCCESS,DONE,"",92,2022-04-22 05:58:26,2022-04-22 05:59:58,"",""
+github:GithubJob:1:4,2022-11-17 06:37:11.409,2022-11-17 06:37:45.836,"{""ConnectionId"":1,""Owner"":""apache"",""Repo"":""incubator-devlake-website""}",_raw_github_api_jobs,751,"",build-and-deploy,github:GithubRun:1:4,SUCCESS,DONE,DEPLOYMENT,191,2022-04-22 06:39:42,2022-04-22 06:42:53,PRODUCTION,""
