+++ plugins/refdiff/e2e/raw_tables/commit_parents.csv
@@ -1,0 +1,6 @@
+commit_sha,parent_commit_sha,created_at,updated_at,_raw_data_params,_raw_data_table,_raw_data_id,_raw_data_remark
+commit_sha1,commit_sha6,2022-11-17 06:37:03.463,2022-11-17 06:37:03.463,"","",0,""
+commit_sha2,commit_sha6,2022-11-17 06:37:03.463,2022-11-17 06:37:03.463,"","",0,""
+commit_sha3,commit_sha6,2022-11-17 06:37:03.463,2022-11-17 06:37:03.463,"","",0,""
+commit_sha4,commit_sha7,2022-11-17 06:37:03.463,2022-11-17 06:37:03.463,"","",0,""
+commit_sha5,commit_sha7,2022-11-17 06:37:03.463,2022-11-17 06:37:03.463,"","",0,""
