+++ plugins/refdiff/e2e/raw_tables/project_mapping.csv
@@ -1,0 +1,4 @@
+project_name,table,row_id
+project1,repos,github:GithubRepo:1:484251804
+project2,repos,github:GithubRepo:1:384111310
+project3,boards,jira:JiraBoard:1:68
