+++ plugins/refdiff/e2e/snapshot_tables/finished_commits_diffs.csv
@@ -1,0 +1,3 @@
+new_commit_sha,old_commit_sha
+commit_sha2,commit_sha1
+commit_sha4,commit_sha2
