+++ plugins/refdiff/refdiff.go
@@ -18,83 +18,12 @@ limitations under the License.
 package main
 
 import (
-	"github.com/apache/incubator-devlake/errors"
-	"github.com/apache/incubator-devlake/plugins/helper"
-	"github.com/apache/incubator-devlake/plugins/refdiff/tasks"
+	"github.com/apache/incubator-devlake/plugins/refdiff/impl"
 	"github.com/apache/incubator-devlake/runner"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	"gorm.io/gorm"
-
-	"github.com/apache/incubator-devlake/plugins/core"
 )
 
-// make sure interface is implemented
-var _ core.PluginMeta = (*RefDiff)(nil)
-var _ core.PluginInit = (*RefDiff)(nil)
-var _ core.PluginTask = (*RefDiff)(nil)
-var _ core.PluginApi = (*RefDiff)(nil)
-
-// PluginEntry is a variable exported for Framework to search and load
-var PluginEntry RefDiff //nolint
-
-type RefDiff struct{}
-
-func (plugin RefDiff) Description() string {
-	return "Calculate commits diff for specified ref pairs based on `commits` and `commit_parents` tables"
-}
-
-func (plugin RefDiff) GetTablesInfo() []core.Tabler {
-	return []core.Tabler{}
-}
-
-func (plugin RefDiff) Init(config *viper.Viper, logger core.Logger, db *gorm.DB) errors.Error {
-	return nil
-}
-
-func (plugin RefDiff) SubTaskMetas() []core.SubTaskMeta {
-	return []core.SubTaskMeta{
-		tasks.CalculateCommitsDiffMeta,
-		tasks.CalculateIssuesDiffMeta,
-		tasks.CalculatePrCherryPickMeta,
-		tasks.CalculateProjectDeploymentCommitsDiffMeta,
-	}
-}
-
-func (plugin RefDiff) PrepareTaskData(taskCtx core.TaskContext, options map[string]interface{}) (interface{}, errors.Error) {
-	var op tasks.RefdiffOptions
-	err := helper.Decode(options, &op, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	db := taskCtx.GetDal()
-	tagsPattern := op.TagsPattern
-	tagsLimit := op.TagsLimit
-	tagsOrder := op.TagsOrder
-
-	rs, err := tasks.CalculateTagPattern(db, tagsPattern, tagsLimit, tagsOrder)
-	if err != nil {
-		return nil, err
-	}
-	op.AllPairs, err = tasks.CalculateCommitPairs(db, op.RepoId, op.Pairs, rs)
-	if err != nil {
-		return nil, err
-	}
-
-	return &tasks.RefdiffTaskData{
-		Options: &op,
-	}, nil
-}
-
-// PkgPath information lost when compiled as plugin(.so)
-func (plugin RefDiff) RootPkgPath() string {
-	return "github.com/apache/incubator-devlake/plugins/refdiff"
-}
-
-func (plugin RefDiff) ApiResources() map[string]map[string]core.ApiResourceHandler {
-	return nil
-}
+var PluginEntry impl.RefDiff
 
 // standalone mode for debugging
 func main() {
